Do not create output directory

+++ test_driver/test_driver.py
@@ -46,6 +46,8 @@ class TestDriver(SingleCrystalTestDriver):
         (MSD) of atoms during the simulation. If the MSD exceeds a given threshold value (msd_threshold), an error is
         raised.
 
+        All output files are written to the "output" directory.
+
         :param temperature_step_fraction:
             Fraction of the target temperature that is used as temperature step for the finite-difference scheme.
             For example, if the target temperature is 300 K and the temperature_step_fraction is 0.1, the temperature
@@ -105,6 +107,7 @@ class TestDriver(SingleCrystalTestDriver):
         :raises KIMTestDriverError:
             If the crystal melts or vaporizes during the simulation.
             If the symmetry of the structure changes.
+            If the output directory "output" does not exist.
         """
         # Set prototype label.
         self.prototype_label = self._get_nominal_crystal_structure_npt()["prototype-label"]["source-value"]
@@ -201,8 +204,9 @@ class TestDriver(SingleCrystalTestDriver):
         assert len(temperatures) == 2 * number_symmetric_temperature_steps + 1
         assert all(t > 0.0 for t in temperatures)
 
-        # Create output directory for all data files and copy over necessary files.
-        os.makedirs("output", exist_ok=True)
+        # Make sure output directory for all data files exists and copy over necessary files.
+        if not os.path.exists("output"):
+            raise KIMTestDriverError("Output directory 'output' does not exist.")
         test_driver_directory = os.path.dirname(os.path.realpath(__file__))
         if os.getcwd() != test_driver_directory:
             shutil.copyfile(os.path.join(test_driver_directory, "npt.lammps"), "npt.lammps")
